Rename YamlMigrationJobConfigurationSwapperTest (#21359)
