+++ kernel/data-pipeline/api/src/main/java/org/apache/shardingsphere/data/pipeline/api/config/job/yaml/YamlConsistencyCheckJobConfiguration.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.data.pipeline.api.config.job.yaml;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
 
 import java.util.Properties;
 
@@ -29,7 +28,6 @@ import java.util.Properties;
  */
 @Getter
 @Setter
-@Slf4j
 @ToString
 public final class YamlConsistencyCheckJobConfiguration implements YamlPipelineJobConfiguration {
